Expose utility functions for loading gltf files
